PARALLAX BACKGROUND WORKS
(Very long-winded explanation ahead)

for the parallax background I initially used an online tutorial, since I wasn't really sure how to do it. However, unlike a lot of the tutorials I used for the basic mechanics, this one didn't really explain much and the movements were kinda janky (though they did work). It also involved a lot of copying and pasting, and I wanted the code to be my own (though coding something like this from scratch a week before this is due was probably not the smartest move).

I ended up deciding to just create a modified version of the camera script which would move at a slightly different speed to the camera. Initially I wanted to find a tutorial for a camera script that made the camera move slightly ahead of the player.

However, when I was browsing forums I found a parallax tutorial which mentioned that parallaxes involved the objects moving at slightly different speeds, so in the end I decided to take the part of the player movement code and put that in the camera's movement position.

I did have a couple of other minor issues with the camera (one of the components was really janky because i accidentally put the camera as the component to follow instead of the player. come to think of it that might be why the first script was so janky. lmao.)

I also added a map border so the player actually dies when they fall off the map. This was just a massive environmental hazard that removed 5 health instead of 1. I might try to make it so that automatically triggers a game over instead of playing the death animation.

also removed herobrine.

(end extremely long-winded rant.)